% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.classification}
\alias{get.classification}
\title{Get Classification Function}
\usage{
get.classification(
  url,
  catalog,
  dataProduct,
  class.id = NULL,
  codeSort = NULL,
  key = NULL,
  limit = NULL,
  offset = NULL
)
}
\arguments{
\item{url}{The base URL of the RDS server}

\item{catalog}{The catalog ID}

\item{dataProduct}{The dataProcuct ID}

\item{class.id}{The classification ID to select}

\item{codeSort}{Specifies how the codes should be sorted, options are ASC or DESC}

\item{key}{API key for dataProducts that require a key.}

\item{offset}{Specifies the starting index of the codes}

\item{imit}{Specifies the number of codes to return}
}
\description{
This function allows you to select a classification based on its ID.
}
\examples{
get.classification("http://localhost:8080/rds/api/catalog/","myCatalog","myDataProduct",class.id="myClass")
get.classification("http://localhost:8080/rds/api/catalog/","myCatalog","myDataProduct",class.id="myClass",limit=10,offset=10)
get.classification("http://localhost:8080/rds/api/catalog/","myCatalog","myDataProduct",class.id="myClass",codeSort="DESC")
}
\keyword{classification}
