% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.rds}
\alias{get.rds}
\title{Get RDS Server}
\usage{
get.rds(apiUrl)
}
\arguments{
\item{apiUrl}{The RDS API URL of the RDS instance that is being accessed.}
}
\description{
This function uses the provided API URL to connect to a RDS server. Connecting to the server will 
validate that this URL is correct and all the subsequent calls for catalogs, data products, metadata,
or data will use this API URL as a base.
}
\examples{
get.rds("http://dev.richdataservices.com/rds/api")
}
\keyword{server}
