% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getCatalog}
\alias{getCatalog}
\title{Get Catalog}
\usage{
getCatalog(server, catalogId)
}
\arguments{
\item{server}{The server to query for the catalog.}

\item{catalogId}{The ID of the desired catalog.}
}
\description{
The getCatalog(server, catalogId) method can be used to get the catalog with the provided ID from the provided rds.server.
}
