% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getCatalogs,rds.server-method}
\alias{getCatalogs,rds.server-method}
\title{Get Catalogs}
\usage{
\S4method{getCatalogs}{rds.server}(server)
}
\arguments{
\item{server}{The server to query for the catalogs.}
}
\description{
The getCatalogs(server) method can be used to get all the catalogs from the provided rds.server.
}
