% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getCatalogs}
\alias{getCatalogs}
\title{Get Catalogs}
\usage{
getCatalogs(server, apiKey = NULL)
}
\arguments{
\item{server}{The server to query for the catalogs.#'}

\item{apiKey}{The user's apiKey to access the API, if the API is not secured this can be NULL.}
}
\description{
The getCatalogs(server) method can be used to get all the catalogs from the provided rds.server.
}
