% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getClassifications}
\alias{getClassifications}
\title{Get Classifications}
\usage{
getClassifications(dataProduct, limit = NULL, offset = NULL)
}
\arguments{
\item{dataProduct}{The rds.dataProduct to query for classifications.}

\item{limit}{Specifies the number of classifications to return.}

\item{offset}{Specifies the starting index of the classifications.}
}
\description{
The getClassifications(dataProduct) method can be used to retrieve the descriptive information about all the classifications that are used in the data product.
}
