% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getDataProduct}
\alias{getDataProduct}
\title{Get Data Product}
\usage{
getDataProduct(catalog, dataProductId)
}
\arguments{
\item{catalog}{The catalog to query for the data product.}

\item{dataProductId}{The ID of the desired data product.}
}
\description{
The getDataProduct(catalog, dataProductId) method can be used to get the data product with the provided ID from the provided rds.catalog.
}
