% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getDataProducts}
\alias{getDataProducts}
\title{Get Data Products}
\usage{
getDataProducts(catalog)
}
\arguments{
\item{catalog}{The catalog to query for the data products.}
}
\description{
The getDataProducts(catalog) method can be used to retrieve all the data products from the provided catalog.
}
