% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getVariable}
\alias{getVariable}
\title{Get Variable Function}
\usage{
getVariable(url, collection, view, var.id = NULL, key = NULL)
}
\arguments{
\item{url}{The base URL of the RDS server}

\item{collection}{The collection ID}

\item{view}{The view ID}

\item{var.id}{The variable ID to select}

\item{key}{API key for views that require a key.}
}
\description{
This function allows you to select the metadata for a single variable based off its ID.
}
\examples{
getVariable("http://localhost:8080/rds/api/catalog/","myCollection","myView","myVariable") 
}
\keyword{variable}

