% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getVariables}
\alias{getVariables}
\title{Get Variables Function}
\usage{
getVariables(url, collection, view, cols = NULL, key = NULL)
}
\arguments{
\item{url}{The base URL of the RDS server}

\item{collection}{The collection ID}

\item{view}{The view ID}

\item{cols}{The columns to select}

\item{key}{API key for views that require a key.}
}
\description{
This function allows you to use the select multiple variables metadata.
}
\examples{
getVariable("http://localhost:8080/rds/api/catalog/","myCollection","myView",cols="var1,var2") 
getVariable("http://localhost:8080/rds/api/catalog/","myCollection","myView",cols="$demographics") 
}
\keyword{variable}

