% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{hasReservedValues,numeric,data.frame-method}
\alias{hasReservedValues,numeric,data.frame-method}
\title{Has Reserved Values}
\usage{
\S4method{hasReservedValues}{numeric,data.frame}(rowNumber, variableDf)
}
\arguments{
\item{rowNumber}{The row number of the data frame that represents the variable in question}

\item{variableDf}{The variable data frame to parse}
}
\description{
A method to check a variable in a DF to see if it has reserved values or not
}
