% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{hasReservedValues}
\alias{hasReservedValues}
\title{Has Reserved Values}
\usage{
hasReservedValues(rowNumber, variableDf)
}
\arguments{
\item{rowNumber}{The row number of the data frame that represents the variable in question}

\item{variableDf}{The variable data frame to check}
}
\description{
A method to check a variable in a DF to see if it has reserved values or not
}
