% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{injectMetadata}
\alias{injectMetadata}
\title{Inject Metadata}
\usage{
injectMetadata(data, metadata)
}
\arguments{
\item{data}{The original data frame from}

\item{collection}{The collection ID}
}
\description{
This function will inject classification categories into the data frame to replace the codes.
}
\examples{
injectMetadata()
}
\keyword{injection,}
\keyword{metadata}

