% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{isClassified,numeric,data.frame-method}
\alias{isClassified,numeric,data.frame-method}
\title{Is Classified}
\usage{
\S4method{isClassified}{numeric,data.frame}(rowNumber, variableDf)
}
\arguments{
\item{rowNumber}{The row number of the data frame that represents the variable in question}

\item{variableDf}{The variable data frame to parse}
}
\description{
A method to check a variable in a DF to see if it has a classification URI set, which indicates that its a classified variable.
}
