% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{isClassified}
\alias{isClassified}
\title{Is Classified}
\usage{
isClassified(rowNumber, variableDf)
}
\arguments{
\item{rowNumber}{The row number of the data frame that represents the variable in question}

\item{variableDf}{The variable data frame to check}
}
\description{
A method to check a variable in a DF to see if it has a classification URI set, which indicates that its a classified variable.
}
