% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{parseVariables,data.frame-method}
\alias{parseVariables,data.frame-method}
\title{Parse Variables}
\usage{
\S4method{parseVariables}{data.frame}(variableDf)
}
\arguments{
\item{variableDf}{The variable data frame to parse}
}
\description{
A method to parse the variable data frames that are returned from the REST call into a list of variable objects.
}
