% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{parseVariables}
\alias{parseVariables}
\title{Parse Variables}
\usage{
parseVariables(variableDf)
}
\arguments{
\item{variableDf}{The variable data frame to parse}
}
\description{
A method to parse the variable data frames that are returned from the REST call into a list of variable objects.
}
