% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{rds.catalog-class}
\alias{rds.catalog-class}
\title{Rich Data Services Catalog}
\description{
A catalog provides basic information about the data products it contains and the data in them. It will also carry the server along to build up the API calls for itself and any data products under it.
}
\section{Slots}{

\describe{
\item{\code{server}}{The server this catalog is a part of.}

\item{\code{id}}{The catalog ID.}

\item{\code{description}}{The catalog description. This may contain embedded HTML.}

\item{\code{label}}{The catalog label.}

\item{\code{name}}{The catalog name.}

\item{\code{dataProductCount}}{The number of data products stored in the catalog}
}}

