% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/methods.R
\docType{class}
\name{rds.catalog}
\alias{rds.catalog}
\alias{getDataProducts}
\alias{getDataProduct}
\title{Rich Data Services Catalog}
\usage{
getDataProducts(catalog)

getDataProduct(catalog, dataProductId)
}
\description{
A catalog provides basic information about the data products it contains and the data in them. It will also carry the server along to build up the API calls for itself and any data products under it.

The getDataProducts(catalog) method can be used to retrieve all the data products from the provided catalog.

The getDataProduct(catalog, dataProductId) method can be used to get the data product with the provided ID from the provided rds.catalog.
}
\section{Slots}{

\describe{
\item{\code{server}}{The server this catalog is a part of.}

\item{\code{id}}{The catalog ID.}

\item{\code{description}}{The catalog description. This may contain embedded HTML.}

\item{\code{label}}{The catalog label.}

\item{\code{name}}{The catalog name.}

\item{\code{dataProductCount}}{The number of data products stored in the catalog}

\item{\code{catalog}}{The catalog to query for the data products.}

\item{\code{catalog}}{The catalog to query for the data product.}

\item{\code{dataProductId}}{The ID of the desired data product.}
}}

