% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{rds.classification-class}
\alias{rds.classification-class}
\title{Rich Data Services Classification}
\description{
A classification contains its ID, the codes and values of the classification, and information about the query run to select the classfication and its codes.
}
\details{
\describe{
   \item{id}{The classification ID.}
   
   \item{codes}{The codes and code values of the classification.}
   
   \item{info}{The information about the query run.}
 }
}

