% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/methods.R
\docType{class}
\name{rds.dataProduct}
\alias{rds.dataProduct}
\alias{getVariables}
\alias{getVariable}
\alias{getClassifications}
\alias{getClassification}
\alias{select}
\alias{tabulate}
\title{Rich Data Services Data Product}
\usage{
getVariables(dataProduct, ..., cols = NULL, collimit = NULL, coloffset = NULL)

getVariable(dataProduct, variableId)

getClassifications(dataProduct, ..., limit = NULL, offset = NULL)

getClassification(dataProduct, classificationId, ..., limit = 1000, offset = 0)

select(
  dataProduct,
  ...,
  limit = NULL,
  offset = NULL,
  cols = NULL,
  colLimit = NULL,
  colOffset = NULL,
  count = FALSE,
  orderby = NULL,
  where = NULL,
  inject = FALSE,
  metadata = FALSE,
  autoPage = TRUE
)

tabulate(
  dataProduct,
  dimensions,
  ...,
  measures = NULL,
  limit = NULL,
  offset = NULL,
  orderby = NULL,
  where = NULL,
  totals = TRUE,
  inject = FALSE,
  metadata = FALSE
)
}
\description{
A data product serves as the basis of which to query data and metadata. It also provides basic information about the data.

The getVariables(dataProduct) method can be used to get the variable summaries of all the with the provided ID from the provided rds.dataProduct.

The getVariable(dataProduct, variableId) method can be used to get the variable with the provided ID from the provided rds.dataProduct.

The getClassifications(dataProduct) method can be used to retrieve the descriptive information about all the classifications that are used in the data product.

The getClassification(dataProduct, classificationId) method can be used to get the classification with the provided ID from the provided rds.dataProduct.

The select(dataProduct) method is used to access the record level data of the data product. In the explorer and through the API the total number of cells is limited to 10000 cells. This is done to keep a small and manageable amount of information going over the network. Be aware that by default the select method will perform numerous calls to build up a complete dataset. If this is not desired remember to set autoPage=FALSE.

The tabulate(dataProduct) method is used to create aggregated tables and perform analysis on a data product.
}
\section{Slots}{

\describe{
\item{\code{catalog}}{The catalog this data product is a part of.}

\item{\code{id}}{The data product ID.}

\item{\code{citation}}{The data product citation. This may contain embedded HTML.}

\item{\code{description}}{The data product description. This may contain embedded HTML.}

\item{\code{label}}{The data product label.}

\item{\code{lastUpdate}}{A timestamp indicating when the data product was last updated.}

\item{\code{name}}{The data product name.}

\item{\code{note}}{Additional information or notes about the data product. This may contain embedded HTML.}

\item{\code{provenance}}{Information about where the data product originated from or who owns it. This may contain embedded HTML.}

\item{\code{restriction}}{Information about any restrictions that may apply to the data of this data product. This may contain embedded HTML.}

\item{\code{dataProduct}}{The rds.dataProduct to query for the variable.}

\item{\code{cols}}{An optional variable query following the RDS column syntax. If left NULL all the variable will be returned.}

\item{\code{collimit}}{An optional limit to the variables returned. If left NULL no limit will be applied.}

\item{\code{coloffset}}{An optional offset to the variables returned. If left NULL no offset will be applied}

\item{\code{dataProduct}}{The rds.dataProduct to query for the variable.}

\item{\code{variableId}}{The ID of the desired variable.}

\item{\code{dataProduct}}{The rds.dataProduct to query for classifications.}

\item{\code{limit}}{Specifies the number of classifications to return.}

\item{\code{offset}}{Specifies the starting index of the classifications.}

\item{\code{dataProduct}}{The rds.dataProduct to query for the classification}

\item{\code{classificationId}}{The ID or URI of the desired classification}

\item{\code{limit}}{Specifies the number of codes to return.}

\item{\code{offset}}{Specifies the starting index of the codes.}

\item{\code{dataProduct}}{The dataProduct whose data is desired.}

\item{\code{limit}}{Specifies the number of records to return.}

\item{\code{offset}}{Specifies the starting index of the records.}

\item{\code{cols}}{The columns to select, these should be specified in the appropriate RDS syntax.}

\item{\code{colLimit}}{Specifies the limit of classifications that should be returned}

\item{\code{colOffset}}{Specifies the starting index of the classifications to be returned}

\item{\code{count}}{Specifies that the total count of records in the dataProduct should be included in the info section.}

\item{\code{orderby}}{Describes how the results should be ordered.}

\item{\code{where}}{Describes how to subset the records based on variable values.}

\item{\code{inject}}{Specifies if metadata should be injected into the data frame. If true and there are classifications available the columns codes will be replaced with code values. Defaults to FALSE}

\item{\code{metadata}}{Specifies if variable metadata should be included in the response.}

\item{\code{autoPage}}{If set to true multiple queries will be sent to the RDS server in order to compile the complete data set.}

\item{\code{dataProduct}}{The dataProduct whose data is being used in the tabulation.}

\item{\code{dimensions}}{The names of the variables that should be used as dimensions.}

\item{\code{measures}}{The variables that should be used as a measure, will be the count by default.}

\item{\code{limit}}{Specifies the number of records to return.}

\item{\code{offset}}{Specifies the starting index of the records.}

\item{\code{orderby}}{Describes how the results should be ordered.}

\item{\code{where}}{Describes the subset of records the tabulation should run on.}

\item{\code{totals}}{Specifies if totals should be included. Totals are used to provide roll up information about the counts of dimensions at different levels.}

\item{\code{inject}}{Specifies if metadata should be injected into the data frame. If true and there are classifications available the columns codes will be replaced with code values. Defaults to FALSE}

\item{\code{metadata}}{Specifies if variable metadata should be included in the response.}
}}

