% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{rds.dataset-class}
\alias{rds.dataset-class}
\title{Rich Data Services Data Set}
\description{
A data set contains the three main sections of an RDS query, metadata, data, and info. The metadata may or may not be present depending on what was requested in the query. If it is it will contain variable and classification metadata if available. The data will contain the actual data values that are returned from the query. The info will contain information about the query ran including the any limits, whether or not more columns or rows are available, and a total row count if available.
}
\section{Slots}{

\describe{
\item{\code{variables}}{A data frame of the variable information of the variables included in the query}

\item{\code{records}}{A data frame of the records returned.}

\item{\code{totals}}{A data frame of the totals returned (tabulation only).}

\item{\code{info}}{A data frame of the query information.}
}}

