% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{rds.metadata-class}
\alias{rds.metadata-class}
\title{Rich Data Services Metadata}
\description{
Metadata consists of variable and classification metadata. The metadata class gives users a way to access the metadata in convenient ways.
}
\details{
\describe{
   \item{json}{The metadata json returned from a query.}
 }
}

