% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{rds.reservedValue-class}
\alias{rds.reservedValue-class}
\title{Rich Data Services reserved values. Reserved values are codes that are set at the variable level (not classification) to identify values with special meaning amongst other non coded values.}
\description{
Rich Data Services reserved values. Reserved values are codes that are set at the variable level (not classification) to identify values with special meaning amongst other non coded values.
}
