% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{rds.server-class}
\alias{rds.server-class}
\title{Rich Data Services Server}
\description{
The RDS server is the basis for the rest of the classes and acts as a reference point for all the API calls to be built from.
}
\section{Slots}{

\describe{
\item{\code{baseUrl}}{The URL to access the base rds server information, swagger, or other server level information. This will also be used to build additional API calls from.}

\item{\code{name}}{The name of the RDS server.}

\item{\code{version}}{The version of the RDS server.}

\item{\code{disclaimer}}{If there is any system wide disclaimer about the data stored in this RDS server there may be a disclaimer that relates to the data. This will be an empty string if there is no disclaimer.}
}}

