% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{rds.tabulate}
\alias{rds.tabulate}
\title{Tabulate}
\usage{
rds.tabulate(
  dataProduct,
  dimensions,
  measures = NULL,
  limit = NULL,
  offset = NULL,
  orderby = NULL,
  weights = NULL,
  where = NULL,
  totals = TRUE,
  inject = FALSE
)
}
\arguments{
\item{dataProduct}{The dataProduct whose data is being used in the tabulation.}

\item{dimensions}{The names of the variables that should be used as dimensions.}

\item{measures}{The variables that should be used as a measure, will be the count by default.}

\item{limit}{Specifies the number of records to return.}

\item{offset}{Specifies the starting index of the records.}

\item{orderby}{Describes how the results should be ordered.}

\item{weights}{Ids of numeric variables to weight the data with. This should only be used for variables that are flagged as weights.}

\item{where}{Describes the subset of records the tabulation should run on.}

\item{totals}{Specifies if totals should be included. Totals are used to provide roll up information about the counts of dimensions at different levels.}

\item{inject}{Specifies if metadata should be injected into the data frame. If true and there are classifications available the columns codes will be replaced with code values. Defaults to FALSE}
}
\description{
The rds.tabulate(dataProduct) method is used to create aggregated tables and perform analysis on a data product.
}
