% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{rds.variable-class}
\alias{rds.variable-class}
\title{Rich Data Services Variable}
\description{
A variable containg the metadata round a column of data. This will provide information about the variable, what it is, the codes that apply to it and summary statistics.
}
\section{Slots}{

\describe{
\item{\code{id}}{The variable ID.}

\item{\code{name}}{The variable name.}

\item{\code{label}}{The variable label.}

\item{\code{description}}{The variable description.}

\item{\code{questionText}}{The question that was asked to get a response. This applies to variables whose data was collected through surveys.}

\item{\code{dataType}}{The data type of the variable. This is a harmonized type that is used accross all data sources regardless of how the data is stored in the back. This is the property that should be evaluated if variable information or data needs to be displayed differently based on type.}

\item{\code{storageType}}{The data type that the variable is stored as in the back end data source. This is specific to the source and cannot be counted on to be the same accross all variables on the RDS server.}

\item{\code{fixedStorageWidth}}{The width of the variable in a fixed file.}

\item{\code{startPosition}}{The start position of the variables data in the context of a fixed file.}

\item{\code{endPosition}}{The end position of the variables data in the context of a fixed file.}

\item{\code{decimals}}{The number of decimal places the variables data may have.}

\item{\code{classificationId}}{If the variable has a classification associated with it its ID will be here. This can be used for display or query purposes, however, the classificationUri will result in faster queries so that should be used for querying.}

\item{\code{classificationUri}}{If the variable has a classification associated with it its URI will be here. This can be used for query purposes.}

\item{\code{index}}{The index of the variable in the data product.}

\item{\code{reference}}{Indicates if the variable is a reference. If true, this variable will not contain the full variable metadata, and the variable metadata should be retrieved from the server if more detail about the variable is desired.}

\item{\code{isDimension}}{Indicates if the variable can be used as a dimension or not in tabulations.}

\item{\code{isMeasure}}{Indicates if the variable can be used as a measure or not in tabulations.}

\item{\code{isRequired}}{Indicates if the variable should always be included in the results or not.}

\item{\code{isWeight}}{Indicates if the variable can be used as a weight.}

\item{\code{statistics}}{A list of rds.statistics objects that potentially contains weighted or non weighted summary statistics of the variables values.}

\item{\code{frequencies}}{A list of rds.frequency objects that potentially contains weighted or non weighted frequencies of the variables values.}
}}

