% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{selectSubset}
\alias{selectSubset}
\title{Select Subset Function}
\usage{
selectSubset(url, collection, view, cols = NULL, where = NULL,
  orderby = NULL, distinct = FALSE, limit = NULL, offset = NULL,
  colLimit = NULL, colOffset = NULL, classLimit = NULL,
  classOffset = NULL, codeLimit = NULL, codeOffset = NULL,
  facts = FALSE, freqs = FALSE, stats = FALSE, count = FALSE,
  varProperties = NULL, key = NULL, metadata = FALSE, inject = FALSE)
}
\arguments{
\item{url}{The base URL of the RDS server}

\item{collection}{The collection ID}

\item{view}{The view ID}

\item{cols}{The columns to select}

\item{where}{Describes how to subset the records}

\item{orderby}{Describes how the results should be ordered}

\item{distinct}{Specifies that only distinct values should be returned}

\item{limit}{Specifies the number of records to return}

\item{offset}{Specifies the starting index of the records}

\item{colLimit}{Specifies the limit of classifications that should be returned}

\item{colOffset}{Specifies the starting index of the classifications to be returned}

\item{classLimit}{Specifies the limit of classifications that should be returned}

\item{classOffset}{Specifies the starting index of the classifications to be returned}

\item{codeLimit}{Specifies the limit of codes that should be returned in each classification}

\item{codeOffset}{Specifies the starting indexes of the codes in each classification}

\item{facts}{Specifies that any facts available should be included in the returned metadata}

\item{freqs}{Specifies that any frequencies available should be included in the returned metadata}

\item{stats}{Specifies that any summary statistics available should be included in the returned}

\item{count}{Specifies that the total count of records in the view should be included in the info section}

\item{varProperties}{Specifies the variable properties to include / exclude from the returned variable metadata. Can be a regex or property names.}

\item{key}{API key for views that require a key.}

\item{metadata}{Specifies that metadata should be returned along with the data. Defaults to FALSE}

\item{inject}{Specifies if metadata should be injected into the data frame. If true and there are classifications available the columns codes will be replaced with code values. Defaults to FALSE}
}
\description{
This function allows users to select a subset of data from RDS.
}
\keyword{select}

