% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.classification}
\alias{get.classification}
\title{Get Classification Function}
\usage{
get.classification(url, collection, view, class.id = NULL, codeSort = NULL,
  key = NULL, limit = NULL, offset = NULL)
}
\arguments{
\item{url}{The base URL of the RDS server}

\item{collection}{The collection ID}

\item{view}{The view ID}

\item{class.id}{The classification ID to select}

\item{codeSort}{Specifies how the codes should be sorted, options are ASC or DESC}

\item{key}{API key for views that require a key.}

\item{limit}{Specifies the number of codes to return}

\item{offset}{Specifies the starting index of the codes}
}
\description{
This function allows you to select a classification based on its ID.
}
\examples{
get.classification("http://localhost:8080/rds/api/catalog/","myCollection","myView",class.id="myClass")
get.classification("http://localhost:8080/rds/api/catalog/","myCollection","myView",class.id="myClass",limit=10,offset=10)
get.classification("http://localhost:8080/rds/api/catalog/","myCollection","myView",class.id="myClass",codeSort="DESC")
}
\keyword{classification}

