% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.classifications}
\alias{get.classifications}
\title{Get Classifications Function}
\usage{
get.classifications(url, collection, view, codeLimit = NULL,
  codeOffset = NULL, codeSort = NULL, key = NULL, limit = NULL,
  offset = NULL, sort = NULL)
}
\arguments{
\item{url}{The base URL of the RDS server}

\item{collection}{The collection ID}

\item{view}{The view ID}

\item{codeLimit}{Specifies the number of codes to return in each classification}

\item{codeOffset}{Specifies the starting index of the codes in each classification}

\item{codeSort}{Specifies how the codes should be sorted, options are ASC or DESC}

\item{key}{API key for views that require a key.}

\item{limit}{Specifies the number of classifications to return}

\item{offset}{Specifies the starting index of the classifications}

\item{sort}{Specifies how to sort the classfications returned. Classifications will be sorted by their ID and can be sorted ASC or DESC}
}
\description{
This function allows you to use the select multiple classifications.
}
\examples{
get.classifications("http://localhost:8080/rds/api/catalog/","myCollection","myView") 
get.classifications("http://localhost:8080/rds/api/catalog/","myCollection","myView",limit=10,offset=10,sort="DESC") 
get.classifications("http://localhost:8080/rds/api/catalog/","myCollection","myView",codeLimit=100,codeSort="DESC")  
}
\keyword{variable}

