% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.rds}
\alias{get.rds}
\title{Get RDS Server}
\usage{
get.rds(host, port = NULL, protocol = "http")
}
\arguments{
\item{host}{The host or domain name of the server}

\item{port}{The port the RDS server is running on if needed. This is NULL by default.}

\item{protocol}{The protocol to use in the call, http by default.}
}
\description{
This function uses the provided host, port, and protocol to build up an RDS server. This will create the base
api paths that the server uses as well as pulling basis information about the server.
}
\examples{
get.rds("http://dev.richdataservices.com")
}
\keyword{server}
