% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.rds}
\alias{get.rds}
\title{Get RDS Server}
\usage{
get.rds(host, port = NULL, protocol = "http", path = "rds")
}
\arguments{
\item{host}{The host or domain name of the server}

\item{port}{The port the RDS server is running on if needed. This is NULL by default.}

\item{protocol}{The protocol to use in the call, http by default.}

\item{path}{The path to the RDS application root. This will be rds by default, if no path is desired (for instances behind a proxy) this can be set to null}
}
\description{
This function uses the provided host, port, protocol, and application path to build up an RDS server. This will create the base
paths that the server uses to build up future requests as well as pulling basic information about the server.
}
\examples{
get.rds("http://dev.richdataservices.com")
get.rds("dev.richdataservices.com", path = "rds")
}
\keyword{server}
