% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.variable}
\alias{get.variable}
\title{Get Variable Function}
\usage{
get.variable(url, collection, view, key = NULL, var.id = NULL)
}
\arguments{
\item{url}{The base URL of the RDS server}

\item{collection}{The collection ID}

\item{view}{The view ID}

\item{key}{API key for views that require a key.}

\item{var.id}{The variable ID to select}
}
\description{
This function allows you to select the metadata for a single variable based off its ID.
}
\examples{
get.variable("http://localhost:8080/rds/api/catalog/","myCollection","myView","myVariable") 
}
\keyword{variable}

