% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getCatalog,rds.server,character-method}
\alias{getCatalog,rds.server,character-method}
\title{Get Catalog}
\usage{
\S4method{getCatalog}{rds.server,character}(server, catalogId, apiKey = NULL)
}
\arguments{
\item{server}{The server to query for the catalog.}

\item{catalogId}{The ID of the desired catalog.}

\item{apiKey}{The user's apiKey to access the API, if the API is not secured this can be NULL.}
}
\description{
The getCatalog(server, catalogId) method can be used to get the catalog with the provided ID from the provided rds.server.
}
