% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getClassification,rds.dataProduct,character-method}
\alias{getClassification,rds.dataProduct,character-method}
\title{Get Classification}
\usage{
\S4method{getClassification}{rds.dataProduct,character}(
  dataProduct,
  classificationId,
  limit = 1000,
  offset = 0,
  apiKey = NULL
)
}
\arguments{
\item{dataProduct}{The rds.dataProduct to query for the classification}

\item{classificationId}{The ID or URI of the desired classification}

\item{limit}{Specifies the number of codes to return.}

\item{offset}{Specifies the starting index of the codes.}

\item{apiKey}{The user's apiKey to access the API, if the API is not secured this can be NULL.}
}
\description{
The getClassification(dataProduct, classificationId) method can be used to get the classification with the provided ID from the provided rds.dataProduct.
}
