% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getDataProducts,rds.catalog-method}
\alias{getDataProducts,rds.catalog-method}
\title{Get Data Products}
\usage{
\S4method{getDataProducts}{rds.catalog}(catalog, apiKey = NULL)
}
\arguments{
\item{catalog}{The catalog to query for the data products.}

\item{apiKey}{The user's apiKey to access the API, if the API is not secured this can be NULL.}
}
\description{
The getDataProducts(catalog) method can be used to retrieve all the data products from the provided catalog.
}
