% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getVariable}
\alias{getVariable}
\title{Get Variable}
\usage{
getVariable(dataProduct, variableId)
}
\arguments{
\item{dataProduct}{The rds.dataProduct to query for the variable.}

\item{variableId}{The ID of the desired variable.}
}
\description{
The getVariable(dataProduct, variableId) method can be used to get the variable with the provided ID from the provided rds.dataProduct.
}
