% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{rds.dataProduct}
\alias{rds.dataProduct}
\title{Rich Data Services Data Product}
\description{
A data product serves as the basis of which to query data and metadata. It also provides basic information about the data.
}
\section{Slots}{

\describe{
\item{\code{catalog}}{The catalog this data product is a part of.}

\item{\code{id}}{The data product ID.}

\item{\code{citation}}{The data product citation. This may contain embedded HTML.}

\item{\code{description}}{The data product description. This may contain embedded HTML.}

\item{\code{label}}{The data product label.}

\item{\code{lastUpdate}}{A timestamp indicating when the data product was last updated.}

\item{\code{name}}{The data product name.}

\item{\code{note}}{Additional information or notes about the data product. This may contain embedded HTML.}

\item{\code{provenance}}{Information about where the data product originated from or who owns it. This may contain embedded HTML.}

\item{\code{restriction}}{Information about any restrictions that may apply to the data of this data product. This may contain embedded HTML.}
}}

