% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{rds.dataset-class}
\alias{rds.dataset-class}
\title{Rich Data Services Data Set}
\description{
A Data Set contains the three main sections of an RDS result, metadata, data, and info. 
The metadata may or may not be present. If it is it will contain variable and classification metadata if available
The data will contain the actual data values that are returned from the query.
The info will contain information about the query ran including the any limits, whether or not more columns or rows are available, and a total row count if available.
}
\details{
\describe{
   \item{metadata}{Variable and classification metadata to accompany the data returned.}

\item{data}{Data returned from a query.}

\item{info}{Information about the query ran.}
 }
}

