% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{select}
\alias{select}
\title{Select}
\usage{
select(
  dataProduct,
  limit = NULL,
  offset = NULL,
  cols = NULL,
  colLimit = NULL,
  colOffset = NULL,
  count = FALSE,
  orderby = NULL,
  where = NULL,
  inject = FALSE,
  autoPage = TRUE
)
}
\arguments{
\item{dataProduct}{The dataProduct whose data is desired.}

\item{limit}{Specifies the number of records to return.}

\item{offset}{Specifies the starting index of the records.}

\item{cols}{The columns to select, these should be specified in the appropriate RDS syntax.}

\item{colLimit}{Specifies the limit of classifications that should be returned}

\item{colOffset}{Specifies the starting index of the classifications to be returned}

\item{count}{Specifies that the total count of records in the dataProduct should be included in the info section.}

\item{orderby}{Describes how the results should be ordered.}

\item{where}{Describes how to subset the records based on variable values.}

\item{inject}{Specifies if metadata should be injected into the data frame. If true and there are classifications available the columns codes will be replaced with code values. Defaults to FALSE}

\item{autoPage}{If set to true multiple queries will be sent to the RDS server in order to compile the complete data set.}
}
\description{
The select(dataProduct) method is used to access the record level data of the data product. In the explorer and through the API the total number of cells is limited to 10000 cells. This is done to keep a small and manageable amount of information going over the network. Be aware that by default the select method will perform numerous calls to build up a complete dataset. If this is not desired remember to set autoPage=FALSE.
}
