% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tabulate}
\alias{tabulate}
\title{Tabulate Function}
\usage{
tabulate(
  url,
  catalog,
  dataProduct,
  dimensions = NULL,
  inject = FALSE,
  key = NULL,
  rowLimit = NULL,
  measures = NULL,
  offset = NULL,
  orderby = NULL,
  where = NULL,
  metadata = FALSE
)
}
\arguments{
\item{url}{The base URL of the server RDS is running on}

\item{catalog}{The catalog ID}

\item{dataProduct}{The dataProduct ID}

\item{dimensions}{The names of the variables that should be used as dimensions}

\item{inject}{Specifies if metadata should be injected into the data frame. If true and there are classifications available the columns codes will be replaced with code values. Defaults to FALSE}

\item{key}{API key for dataProducts that require a key.}

\item{rowLimit}{Specifies the number of records to return}

\item{measures}{The variables that should be used as a measure, will be the count by default}

\item{offset}{Specifies the starting index of the records}

\item{orderby}{Describes how the results should be ordered}

\item{where}{Describes the subset of records the tabulation should run on}
}
\description{
This function allows you to use the tabulate function of RDS.
}
\examples{
tabulate("http://localhost:8080/rds/api/catalog/","myCatalog","myDataProduct",dimensions="var1,var2") 
tabulate("http://localhost:8080/rds/api/catalog/","myCatalog","myDataProduct",dimensions="var1,var2",measures="AVG(var3)")
tabulate("http://localhost:8080/rds/api/catalog/","myCatalog","myDataProduct",dimensions="var1,var2",measures="AVG(var3)",where="var1!=5",inject=TRUE) 
}
\keyword{tabulate}
