% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tabulate}
\alias{tabulate}
\title{Tabulate Function}
\usage{
tabulate(url, collection, view, dimensions = NULL, measures = NULL,
  orderby = NULL, where = NULL, limit = NULL, offset = NULL,
  count = FALSE, key = NULL, inject = FALSE)
}
\arguments{
\item{url}{The base URL of the server RDS is running on}

\item{collection}{The collection ID}

\item{view}{The view ID}

\item{orderby}{Describes how the results should be ordered}

\item{where}{Describes how to subset the records}

\item{limit}{Specifies the number of records to return}

\item{offset}{Specifies the starting index of the records}

\item{count}{Specifies that the total count of records in the view should be included in the info section}

\item{key}{API key for views that require a key.}

\item{inject}{Specifies if metadata should be injected into the data frame. If true and there are classifications available the columns codes will be replaced with code values. Defaults to FALSE}

\item{cols}{The columns to select}

\item{distinct}{Specifies that only distinct values should be returned}

\item{classLimit}{Specifies the limit of classifications that should be returned}

\item{classOffset}{Specifies the starting index of the classifications to be returned}

\item{codeLimit}{Specifies the limit of codes that should be returned in each classification}

\item{codeOffset}{Specifies the starting indexes of the codes in each classification}

\item{facts}{Specifies that any facts available should be included in the returned metadata}

\item{freqs}{Specifies that any frequencies available should be included in the returned metadata}

\item{stats}{Specifies that any summary statistics available should be included in the returned}
}
\description{
This function allows you to use the tabulate function of RDS.
}
\examples{
tabulate("http://localhost:8080/rds/api/catalog/","myCollection","myView",dimensions="var1,var2") 
tabulate("http://localhost:8080/rds/api/catalog/","myCollection","myView",dimensions="var1,var2",measures="AVG(var3)")
tabulate("http://localhost:8080/rds/api/catalog/","myCollection","myView",dimensions="var1,var2",measures="AVG(var3)",where="var1!=5",inject=TRUE) 
}
\keyword{tabulate}

